Unused casts
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/batik/trunk@1831630 13f79535-47bb-0310-9956-ffa450edef68

+++ batik-awt-util/src/main/java/org/apache/batik/ext/awt/g2d/AbstractGraphics2D.java
@@ -605,7 +605,7 @@ public abstract class AbstractGraphics2D extends Graphics2D implements Cloneable
      * @see         java.awt.Graphics#drawChars
      */
     public void drawString(String str, int x, int y){
-        drawString(str, (float)x, (float)y);
+        drawString(str, (float)x, y);
     }
 
 
@@ -623,7 +623,7 @@ public abstract class AbstractGraphics2D extends Graphics2D implements Cloneable
      */
     public void drawString(AttributedCharacterIterator iterator,
                            int x, int y){
-        drawString(iterator, (float)x, (float)y);
+        drawString(iterator, (float)x, y);
     }
 
     /**
